Differentiate between primitives and profile types
The idea is to get the code structure closer to how types are organized
within the data type repository. Every property of a profile is now seen
as a type, which wraps the actual type(s) behind it.
The structure is not completely fulfilled as not every primitive gets
wrapped as a type but are still used in the CreateComponent. But the
structure of the modules is already there, so future changes may be
smaller.

Examples for primitive types are Pid, String and DateTimeHandle.
Examples for types within profiles are CreationDate and
ModificationDate.